Split WrapUtil into CombineWrapUtil and FixWrapUtil
