+++ src/main/java/tudelft/ti2806/pl3/data/wrapper/util/VerticalWrapUtil.java
@@ -54,13 +54,13 @@ public final class VerticalWrapUtil {
 	private static List<Wrapper> combineNodes(List<Wrapper> nodes) {
 		Map<Integer, Wrapper> nonWrappedNodes = new HashMap<>(nodes.size());
 		List<Integer> nonWrappedNodesOrder = new ArrayList<>(nodes.size());
-		WrapUtil.fillNonWrappedCollections(nodes, nonWrappedNodes, nonWrappedNodesOrder);
+		CombineWrapUtil.fillNonWrappedCollections(nodes, nonWrappedNodes, nonWrappedNodesOrder);
 		List<CombineWrapper> combinedNodes = generateCombinedNodes(nodes, nonWrappedNodes);
 		if (combinedNodes.size() == 0) {
 			return null;
 		}
-		List<Wrapper> result = WrapUtil.collectNonWrappedNodes(nonWrappedNodes, nonWrappedNodesOrder);
-		return WrapUtil.wrapAndReconnect(result, combinedNodes);
+		List<Wrapper> result = CombineWrapUtil.collectNonWrappedNodes(nonWrappedNodes, nonWrappedNodesOrder);
+		return CombineWrapUtil.wrapAndReconnect(result, combinedNodes);
 	}
 	
 	/**
